admin finished
You can now set different roles for users.

There's a general user and an admin

If you log in as a user, you're directed to a general user view.

General users are unable to visit admin protected views. To maintain security, this check is made on both the client and server side.

Admin's are automatically directed to their specific view if specified in the database. Admins are hardcoded and can't be generated by users.