Dynamic Dashboard Finished
The user dashboard now has "completed" and "remaining" columns that dynamically update in accordance with document data stored for a particular user.

Documents that are stored as null in the database are initially in the "remaining" column. When a user fills out a document it's moved to the "completed" column and can be viewed.

No new packages installed